pyproject.toml
[tool.poetry]
name = "kit"
version = "0.1.0"
description = ""
authors = ["omid <omid9h@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
pydantic-settings = "^2.0.3"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
django-filter = "^23.2"
django-cors-headers = "^4.2.0"
inject = "^5.0.0"
pillow = "^10.0.0"
psycopg2-binary = "^2.9.7"
django-redis = "^5.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"
autoflake = "^2.2.0"
django-debug-toolbar = "^4.2.0"
faker = "^19.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
# override black's default line-length
line-length = 120
exclude = '(\.eggs|\.git|\.hg|\.mypy_cache|\.nox|\.tox|\.venv|_build|buck-out|build|dist)'

[tool.isort]
# make it compatible with black
profile = "black"
